Move setupLogging bridge into @beakerstack/logger.
Consolidate duplicated web/mobile setupLogging into the logger package, colocate tests there, and address review nits (TODO for telemetry wiring, metadata-only email dev logs).
