+++ apps/mobile/index.js
@@ -1,10 +1,11 @@
 import { registerRootComponent } from 'expo';
 import { initObservability } from '@beakerstack/observability/native';
 import { beakerstackObservabilityConfig } from './src/config/observability';
-import { setupLogging } from './src/setupLogging';
+import { setupLogging } from '@beakerstack/logger';
 import App from './App';
 
 void initObservability(beakerstackObservabilityConfig);
+// TODO: pass observability handle after init — setupLogging(telemetry)
 setupLogging();
 
 registerRootComponent(App);
